requirements.txt
dj-database-url==0.5.0
Django==2.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
